Optimization and bug fix in 2D
Optimization of the RandomFieldSampler class introducing a cache for the deterministic part of the KL expansion.
Bug fix in the FVSolver class: when working on the borders of the mesh grid, the values of k must be taken exactly on the border, not at the centre of the mesh square.